+++ src/index.js
@@ -13,6 +13,7 @@ export { default as TdInputNumber } from './td-input-number';
 export { default as toast } from './toast';
 export { default as DownloadBlob } from './download-blob';
 export { default as Reject } from './reject';
+export { default as Magic } from './magic';
 
 // 高耦合组件
 export { default as localConfig } from './local-config';